feat: add 게임 전체 조회 , 검색 코드 통합 /games
1. 게임 전체 조회 , 검색 코드 통합 /games 로 통합하여 사용
